requirements.txt
fastapi==0.121.0
bitsandbytes==0.48.2
peft==0.17.1
huggingface-hub==0.36.0
loguru==0.7.3
python-dotenv==1.2.1
transformers==4.57.1
uvicorn==0.38.0
vllm==0.11.0
xformers==0.0.32.post1
